Fix MSVC warning.

+++ gamepad.c
@@ -130,8 +130,8 @@ void GamepadSetRumble(GAMEPAD_DEVICE gamepad, float left, float right) {
 	if ((STATE[gamepad].flags & FLAG_RUMBLE) != 0) {
 		XINPUT_VIBRATION vib;
 		ZeroMemory(&vib, sizeof(vib));
-		vib.wLeftMotorSpeed = left * 65535;
-		vib.wRightMotorSpeed = right * 65535;
+		vib.wLeftMotorSpeed = (WORD)(left * 65535);
+		vib.wRightMotorSpeed = (WORD)(right * 65535);
 		XInputSetState(gamepad, &vib);
 	}
 }
